Move tests to src folder
