+++ src/tests/datasets/test_samsum_datasets.py
@@ -2,17 +2,17 @@
 # This software may be used and distributed according to the terms of the Llama 2 Community License Agreement.
 
 import pytest
+from functools import partial
 from unittest.mock import patch
-
 
 EXPECTED_RESULTS = {
     "meta-llama/Llama-2-7b-hf":{
-        "label": 1152,
-        "pos": 31,
+        "label": 8432,
+        "pos": 242,
     },
     "meta-llama/Meta-Llama-3-8B":{
-        "label": 40,
-        "pos": 26,
+        "label": 2250,
+        "pos": 211,
     },
 }
 
@@ -22,10 +22,11 @@ EXPECTED_RESULTS = {
 @patch('llama_recipes.finetuning.LlamaForCausalLM.from_pretrained')
 @patch('llama_recipes.finetuning.optim.AdamW')
 @patch('llama_recipes.finetuning.StepLR')
-def test_grammar_dataset(step_lr, optimizer, get_model, tokenizer, train, setup_tokenizer, llama_version):
+def test_samsum_dataset(step_lr, optimizer, get_model, tokenizer, train, mocker, setup_tokenizer, llama_version):
     from llama_recipes.finetuning import main
 
     setup_tokenizer(tokenizer)
+    get_model.return_value.get_input_embeddings.return_value.weight.shape = [32000 if "Llama-2" in llama_version else 128256]
 
     BATCH_SIZE = 8
     kwargs = {
@@ -33,7 +34,7 @@ def test_grammar_dataset(step_lr, optimizer, get_model, tokenizer, train, setup_
         "batch_size_training": BATCH_SIZE,
         "val_batch_size": 1,
         "use_peft": False,
-        "dataset": "grammar_dataset",
+        "dataset": "samsum_dataset",
         "batching_strategy": "padding",
         }
 
@@ -44,9 +45,10 @@ def test_grammar_dataset(step_lr, optimizer, get_model, tokenizer, train, setup_
     args, kwargs = train.call_args
     train_dataloader = args[1]
     eval_dataloader = args[2]
+    token = args[3]
 
-    VAL_SAMPLES = 2988
-    TRAIN_SAMPLES = 13016
+    VAL_SAMPLES = 818
+    TRAIN_SAMPLES = 14732
 
     assert len(train_dataloader) == TRAIN_SAMPLES // BATCH_SIZE
     assert len(eval_dataloader) == VAL_SAMPLES
@@ -60,7 +62,6 @@ def test_grammar_dataset(step_lr, optimizer, get_model, tokenizer, train, setup_
     assert batch["labels"][0][EXPECTED_RESULTS[llama_version]["pos"]-1] == -100
     assert batch["labels"][0][EXPECTED_RESULTS[llama_version]["pos"]] == EXPECTED_RESULTS[llama_version]["label"]
 
-    token = args[3]
     assert batch["input_ids"][0][0] == token.bos_token_id
     assert batch["labels"][0][-1] == token.eos_token_id
     assert batch["input_ids"][0][-1] == token.eos_token_id
